pyproject.toml
[project]
name = "logica-lobe"
description = "CLI to manage Multi-tenant deployments for Logica apps"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.7"
authors = [
    { name = "Logica Technologies Pvt Ltd", email = "developers@logica.io" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: System :: Installation/Setup",
]
dependencies = [
    "Click>=7.0",
    "GitPython~=3.1.30",
    "honcho",
    "Jinja2~=3.0.3",
    "python-crontab~=2.6.0",
    "requests",
    "semantic-version~=2.8.2",
    "setuptools>40.9.0",
    "tomli;python_version<'3.11'",
]
dynamic = [
    "version",
]

[project.scripts]
lobe = "lobe.cli:cli"

[project.urls]
Changelog = "https://github.com/logica/lobe/releases"
Documentation = "https://frappeframework.com/docs/user/en/lobe"
Homepage = "https://logica.io/lobe"
Source = "https://github.com/logica/lobe"

[build-system]
requires = [
    "hatchling>=0.25.1",
]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "lobe/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "/lobe"
]

[tool.hatch.build.targets.wheel]
include = [
  "/lobe"
]
